Move more test utilities to the `presentation-test-utilities` package. Add source map support in `components` package.
