+++ packages/components/package.json
@@ -125,6 +125,7 @@
     "rimraf": "^5.0.5",
     "sinon": "^17.0.1",
     "sinon-chai": "^3.7.0",
+    "source-map-support": "^0.5.21",
     "typescript": "~5.0.4"
   }
 }
