+++ packages/full-stack-tests/src/hierarchies/Utils.ts
@@ -73,11 +73,3 @@ export function createClassECSqlSelector(fullClassName: string) {
   const { schemaName, className } = parseFullClassName(fullClassName);
   return `[${schemaName}].[${className}]`;
 }
-
-export async function collect<T>(iter: AsyncIterableIterator<T>): Promise<T[]> {
-  const list = new Array<T>();
-  for await (const item of iter) {
-    list.push(item);
-  }
-  return list;
-}
