+++ packages/hierarchies-react/package.json
@@ -80,6 +80,7 @@
     "global-jsdom": "^24.0.0",
     "mocha": "^10.3.0",
     "nyc": "^15.1.0",
+    "presentation-test-utilities": "workspace:^",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "rimraf": "^5.0.5",
