+++ packages/hierarchies-react/src/test/UseTree.test.tsx
@@ -4,6 +4,7 @@
  *--------------------------------------------------------------------------------------------*/
 
 import { expect } from "chai";
+import { createAsyncIterator, throwingAsyncIterator } from "presentation-test-utilities";
 import { PropsWithChildren } from "react";
 import { act } from "react-dom/test-utils";
 import sinon from "sinon";
@@ -12,16 +13,7 @@ import { createStorage, Selectables, StorageSelectionChangeEventArgs, StorageSel
 import { PresentationHierarchyNode, PresentationInfoNode, UnifiedSelectionProvider } from "../presentation-hierarchies-react";
 import { createNodeId } from "../presentation-hierarchies-react/internal/Utils";
 import { useTree, useUnifiedSelectionTree } from "../presentation-hierarchies-react/UseTree";
-import {
-  cleanup,
-  createAsyncIterator,
-  createStub,
-  createTestGroupingNode,
-  createTestHierarchyNode,
-  renderHook,
-  throwingAsyncIterator,
-  waitFor,
-} from "./TestUtils";
+import { cleanup, createStub, createTestGroupingNode, createTestHierarchyNode, renderHook, waitFor } from "./TestUtils";
 
 describe("useTree", () => {
   const hierarchyProvider = {
