+++ packages/hierarchies/src/test/internal/LimitingECSqlQueryExecutor.test.ts
@@ -4,11 +4,11 @@
  *--------------------------------------------------------------------------------------------*/
 
 import { expect } from "chai";
+import { collect, createAsyncIterator } from "presentation-test-utilities";
 import sinon from "sinon";
 import { IECSqlQueryExecutor, trimWhitespace } from "@itwin/presentation-shared";
 import { RowsLimitExceededError } from "../../hierarchies/HierarchyErrors";
 import { createLimitingECSqlQueryExecutor } from "../../hierarchies/LimitingECSqlQueryExecutor";
-import { collect, createFakeQueryReader } from "../Utils";
 
 describe("createLimitingECSqlQueryExecutor", () => {
   const baseExecutor = {
@@ -21,32 +21,32 @@ describe("createLimitingECSqlQueryExecutor", () => {
 
   it("returns base executor's result", async () => {
     const row = [1, 2, 3];
-    baseExecutor.createQueryReader.returns(createFakeQueryReader([row]));
+    baseExecutor.createQueryReader.returns(createAsyncIterator([row]));
     const limitingExecutor = createLimitingECSqlQueryExecutor(baseExecutor, 1);
     const rows = await collect(limitingExecutor.createQueryReader({ ecsql: "query" }));
     expect(rows).to.deep.eq([row]);
   });
 
   it("throws when base executor returns more rows than the limit", async () => {
-    baseExecutor.createQueryReader.returns(createFakeQueryReader([{}, {}]));
+    baseExecutor.createQueryReader.returns(createAsyncIterator([{}, {}]));
     const limitingExecutor = createLimitingECSqlQueryExecutor(baseExecutor, 1);
     await expect(collect(limitingExecutor.createQueryReader({ ecsql: "query" }))).to.eventually.be.rejectedWith(RowsLimitExceededError);
   });
 
   it(`calls base executor with original query when limit is "unbounded"`, async () => {
-    baseExecutor.createQueryReader.returns(createFakeQueryReader([{}]));
+    baseExecutor.createQueryReader.returns(createAsyncIterator([{}]));
     await collect(createLimitingECSqlQueryExecutor(baseExecutor, "unbounded").createQueryReader({ ecsql: "query" }));
     expect(baseExecutor.createQueryReader).to.be.calledOnceWith("query");
   });
 
   it(`calls base executor with added CTEs`, async () => {
-    baseExecutor.createQueryReader.returns(createFakeQueryReader([{}]));
+    baseExecutor.createQueryReader.returns(createAsyncIterator([{}]));
     await collect(createLimitingECSqlQueryExecutor(baseExecutor, "unbounded").createQueryReader({ ecsql: "query", ctes: ["cte1", "cte2"] }));
     expect(baseExecutor.createQueryReader).to.be.calledOnceWith("WITH RECURSIVE cte1, cte2 query");
   });
 
   it(`calls base executor with added limits`, async () => {
-    baseExecutor.createQueryReader.returns(createFakeQueryReader([{}]));
+    baseExecutor.createQueryReader.returns(createAsyncIterator([{}]));
     await collect(createLimitingECSqlQueryExecutor(baseExecutor, 1).createQueryReader({ ecsql: "query" }));
     expect(baseExecutor.createQueryReader).to.be.calledOnceWith(
       sinon.match((ecsql) => trimWhitespace(ecsql) === trimWhitespace("SELECT * FROM (query) LIMIT 2")),
@@ -54,7 +54,7 @@ describe("createLimitingECSqlQueryExecutor", () => {
   });
 
   it(`calls base executor with added CTEs and limits`, async () => {
-    baseExecutor.createQueryReader.returns(createFakeQueryReader([{}]));
+    baseExecutor.createQueryReader.returns(createAsyncIterator([{}]));
     await collect(createLimitingECSqlQueryExecutor(baseExecutor, 1).createQueryReader({ ecsql: "query", ctes: ["cte1", "cte2"] }));
     expect(baseExecutor.createQueryReader).to.be.calledOnceWith(
       sinon.match((ecsql) => trimWhitespace(ecsql) === trimWhitespace("WITH RECURSIVE cte1, cte2 SELECT * FROM (query) LIMIT 2")),
