+++ packages/test-utilities/package.json
@@ -16,7 +16,8 @@
   "dependencies": {
     "@itwin/core-bentley": "^4.5.1",
     "@itwin/core-common": "^4.5.1",
-    "@itwin/core-geometry": "^4.5.1"
+    "@itwin/core-geometry": "^4.5.1",
+    "rxjs": "^7.8.1"
   },
   "devDependencies": {
     "@itwin/eslint-plugin": "4.0.0-dev.48",
