+++ packages/test-utilities/src/test-utilities/AsyncIterators.ts
@@ -3,8 +3,12 @@
  * See LICENSE.md in the project root for license terms and full copyright notice.
  *--------------------------------------------------------------------------------------------*/
 
-export * from "./test-utilities/AsyncIterators";
-export * from "./test-utilities/IModelUtils";
-export * from "./test-utilities/Observables";
-export * from "./test-utilities/ResolvablePromise";
-export * from "./test-utilities/WaitFor";
+export async function* createAsyncIterator<T>(values: T[]): AsyncIterableIterator<T> {
+  for (const value of values) {
+    yield value;
+  }
+}
+
+export async function* throwingAsyncIterator(error: Error): AsyncIterableIterator<never> {
+  throw error;
+}
